requirements.txt
celery==5.3.6
django=5.0.2
djangorestframework=3.14.0
markdown=3.5.2